final_project/requirements.txt
Flask>=2.2.2
Flask-WTF==1.1.1
ibm-watson==6.1.0
python-dotenv==1.0.0
pyJWT==2.6.0
